Migrate to CLI Args
Change the tool to use CLI Args instead of environment variables
