--- main_test.go
@@ -1,3 +1,0 @@
-package main
-
-
